general fix

+++ back-end/database/migrations/2019_04_24_163914_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             //$table->uuid('order_id');
             $table->bigInteger('address_id')->unsigned();
             $table->bigInteger('referral_id')->unsigned()->nullable();
-            $table->integer('shipper_id')->unsigned();
+            $table->bigInteger('shipper_id')->unsigned();
             $table->string('_token',20);
             $table->enum('status',['pending','active','shipped','delivered','canceled','returned'])->default('pending');
             $table->string('status_message',150)->nullable();
